Move pseudolikelihood substitution-matrix reference to tests
The generic log_pseudolikelihood_sites/log_pseudolikelihood_exact
methods, substitution_matrix_sites, substitution_matrix_exhaustive,
their _2states helpers, and _with_leading_dims were unreachable through
normal dispatch: every visible layer with a substitution-matrix method
(Binary, Spin, Potts, PottsGumbel) also has a specialized fast path that
shadows the generic methods, and any other visible type hits a
MethodError either way. The only callers were tests, which needed
invoke() to reach the generic path.

Move the reference implementation into a test-only module,
test/pseudolikelihood_reference.jl, and use it directly as the oracle in
test/pseudolikelihood.jl, dropping the invoke() calls that targeted the
removed generic methods. The invoke() calls that pin the non-Gaussian
Potts fast path remain, since those methods stay in src.

Start the 5.8.3-DEV development cycle with a changelog entry, since the
removed functions were importable by name even though non-exported.

Closes #172.
