+++ Project.toml
@@ -1,7 +1,7 @@
 name = "RestrictedBoltzmannMachines"
 uuid = "12e6b396-7db5-4506-8cb6-664a4fe1e50e"
 authors = ["Jorge FdCD <jorge.fdcd@icloud.com>"]
-version = "5.8.2"
+version = "5.8.3-DEV"
 
 [workspace]
 projects = ["test", "docs", "notebooks", "repl"]
